+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - set up release-it ([30a37c1](https://github.com/adamhl8/configs/commit/30a37c1242edc364e071da9a9d62a1f3ad1208f0))
 
-- *(git-cliff)* set filter_commits to true ([cd9c7f2](https://github.com/adamhl8/configs/commit/cd9c7f2178c1ec77859bc1a9596184c9eda5f462))
+- _(git-cliff)_ set filter_commits to true ([cd9c7f2](https://github.com/adamhl8/configs/commit/cd9c7f2178c1ec77859bc1a9596184c9eda5f462))
+
 ## [0.22.0] - 2026-06-29
 
 ### 🚀 Features